Move sandbox branch retrieval to PullDialogContainer #5610
