+++ ui/app/src/components/GitManagement/PullDialog/utils.ts
@@ -20,11 +20,8 @@ import ApiResponse from '../../../models/ApiResponse';
 import { PullResponse } from '../../../services/repositories';
 
 export interface PullFromRemoteBaseProps {
-  sandboxBranch: string;
-  sandboxBranchError: ApiResponse;
   remoteName: string;
   mergeStrategies: MergeStrategy[];
-  onFetchSandboxBranch(): void;
 }
 
 export interface PullFromRemoteDialogProps extends PullFromRemoteBaseProps, EnhancedDialogProps {
